panel/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
grpcio>=1.68.0
grpcio-tools>=1.68.0
protobuf>=5.28.0
pydantic>=2.9.0
python-dotenv>=1.0.0
